clean norme and add value in enum to accept range to -27

+++ libft/string/ft_strmapi.c
@@ -6,22 +6,22 @@
 /*   By: fkante <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/11 15:29:42 by fkante            #+#    #+#             */
-/*   Updated: 2020/05/18 23:36:25 by francis          ###   ########.fr       */
+/*   Updated: 2020/06/09 11:08:16 by amartino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strmapi(char const *s, char (*f)(uint32_t, char))
+char	*ft_strmapi(char const *s, char (*f)(size_t, char))
 {
-	uint32_t		i;
-	uint32_t		len;
-	char			*tmp;
+	size_t		i;
+	size_t		len;
+	char		*tmp;
 
 	if (s == NULL || f == NULL)
 		return (NULL);
 	i = 0;
-	len = (uint32_t)ft_strlen(s);
+	len = ft_strlen(s);
 	if (!(tmp = (char *)malloc(len + 1)))
 		return (NULL);
 	while (i < len)
